fixup! JAMES-2393 Introduce dedicated modules for event sourcing
