+++ event-sourcing/event-sourcing-cassandra/pom.xml
@@ -57,6 +57,11 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>apache-james-eventsourcing-core</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>apache-james-eventsourcing-core</artifactId>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
