+++ event-sourcing/event-sourcing-memory/pom.xml
@@ -47,6 +47,11 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>apache-james-eventsourcing-core</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>apache-james-eventsourcing-core</artifactId>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
