+++ mailbox/plugin/quota-mailing/pom.xml
@@ -40,6 +40,11 @@
         </dependency>
         <dependency>
             <groupId>org.apache.james</groupId>
+            <artifactId>apache-james-eventsourcing-core</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.james</groupId>
             <artifactId>apache-james-eventsourcing-api</artifactId>
             <type>test-jar</type>
             <scope>test</scope>
